Extract turn_constants.rs from turn.rs (parent #680)
Move the five module-level `pub(super) const` items out of `turn.rs`
into a focused sibling module so `turn.rs` keeps shrinking toward
responsibility-focused units (parent meta-issue #680).

Five shared constants extracted:

- `LOG_ARGS_MAX_CHARS` (= 200) — trace-log argument truncation cap
- `PLAN_REPEATED_EXPLORATION_BLOCK_THRESHOLD` (= 2) — actor-loop
  plan-mode repeated-exploration detector threshold
- `TASK_CONTRACT_VERIFIER_ATTEMPT_LIMIT` (= 3) — verifier diagnostic
  attempt budget
- `TASK_CONTRACT_VERIFIER_REPAIR_ATTEMPT_LIMIT` (= 6) — verifier
  repair attempt budget
- `USER_INTERRUPT_ERROR` (= "__anvil_user_interrupt__") — sentinel
  error string for user-interrupt early returns

Values byte-for-byte unchanged; only the namespace moved from
`super::turn::X` to `super::turn_constants::X`.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `recovery_targets.rs`, `streaming_reply.rs`, `actor_loop_flow.rs`,
`reply_retry.rs`, `verifier_orchestration.rs`, and `progress_tests.rs`
(import path swap, no behavioral change).

LOC delta:
- turn.rs: 68 → 57 (-11)
- new module: +32

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
